testMockUserRoles moved to json file
