+++ src/components/Budget/BudgetCategory/__tests__/BudgetCategory.test.tsx
@@ -11,7 +11,7 @@ import { budgetCategoryFormInputFields } from "../inputFields.json";
 import { commonFormTestUtil } from "../../../../utils/commonFormTest.util";
 import { IBudgetCategory } from "../../../../models/budget";
 import { FORM_ACTIONS } from "../../../../models/budget/constants";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 const handleClose = jest.fn();
