+++ src/components/Budget/BudgetLineitem/__test__/BudgetLineItem.test.tsx
@@ -32,7 +32,7 @@ import { act } from "react-dom/test-utils";
 import { budgetLineitemFormInputFields, budgetLineitemFormSelectFields } from "../inputFields.json";
 import { commonFormTestUtil } from "../../../../utils/commonFormTest.util";
 import { fireEvent, wait } from "@testing-library/dom";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 const handleClose = jest.fn();
