+++ src/components/Budget/BudgetTarget/__test__/BudgetTarget.test.tsx
@@ -19,7 +19,7 @@ import { budgetTargetFormInputFields, budgetTargetFormSelectFields } from "../in
 import { commonFormTestUtil } from "../../../../utils/commonFormTest.util";
 import { IBudgetTargetForm } from "../../../../models/budget/budgetForm";
 import { fireEvent, wait } from "@testing-library/dom";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 const handleClose = jest.fn();
