+++ src/components/Dasboard/ProjectName/__test__/ProjectName.test.tsx
@@ -7,7 +7,7 @@ import { act } from "react-dom/test-utils";
 import { DashboardProvider } from "../../../../contexts/dashboardContext";
 import { NotificationProvider } from "../../../../contexts/notificationContext";
 import { projectDetails } from "../../../../utils/testMock.json";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 const getProjectMock = {
