+++ src/components/Deliverable/__test__/DeliverableTrackline.test.tsx
@@ -25,7 +25,7 @@ import {
 	projectDonorMock,
 	financialYearListMock,
 } from "../../../utils/testMock.json";
-import { mockUserRoles } from "../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../graphql/User/query";
 let createDeliverableTracklineMutation = false;
 const mocks = [
