+++ src/components/Deliverable/__test__/DeliverableUnit.test.tsx
@@ -10,7 +10,7 @@ import { DashboardProvider } from "../../../contexts/dashboardContext";
 import { NotificationProvider } from "../../../contexts/notificationContext";
 import { organizationDetail } from "../../../utils/testMock.json";
 import { deliverableCategoryMock } from "./testHelp";
-import { mockUserRoles } from "../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../graphql/User/query";
 
 const mocks = [
