+++ src/components/Table/Deliverable/__test__/DeliverableTables.test.tsx
@@ -27,7 +27,7 @@ import { GET_DELIVERABLE_ORG_CATEGORY } from "../../../../graphql/Deliverable/ca
 import { organizationDetail, financialYearListMock } from "../../../../utils/testMock.json";
 import { GET_ANNUAL_YEARS, GET_FINANCIAL_YEARS } from "../../../../graphql";
 import { annualYearListMock } from "../../../Impact/__test__/testHelp";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 let intialFormValue = {
