+++ src/components/Table/DeliverableCategoryTable/__test__/DeliverableCategoryTable.test.tsx
@@ -31,7 +31,7 @@ import {
 import { GET_DELIVERABLE_UNIT_PROJECT_COUNT } from "../../../../graphql/Deliverable/unit";
 import { commonFormTestUtil } from "../../../../utils/commonFormTest.util";
 import { deliverableUnitInputFields } from "../../../../pages/settings/DeliverableMaster/inputFields.json";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 let table: any;
