+++ src/components/Table/Donor/__test__/DonorTable.test.tsx
@@ -14,7 +14,7 @@ import {
 import { GET_COUNTRY_LIST } from "../../../../graphql";
 import DonorTable from "../DonorTable";
 import { donorTableHeading } from "../../constants";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 let table: any;
