+++ src/components/Table/ImpactUnitTable/__test__/impactUnitTable.test.tsx
@@ -28,7 +28,7 @@ import {
 import { GET_IMPACT_CATEGORY_PROJECT_COUNT } from "../../../../graphql/Impact/category";
 import { commonFormTestUtil } from "../../../../utils/commonFormTest.util";
 import { impactCategoryInputFields } from "../../../../pages/settings/ImpactMaster/inputFields.json";
-import { mockUserRoles } from "../../../../utils/testMockUserRoles";
+import { mockUserRoles } from "../../../../utils/testMockUserRoles.json";
 import { GET_USER_ROLES } from "../../../../graphql/User/query";
 
 let table: any;
