--- src/utils/testMockUserRoles.ts
@@ -1,612 +1,0 @@
-export const mockUserRoles = {
-	role: {
-		id: "1",
-		name: "Authenticated",
-		organization: {
-			id: "3",
-			name: "sherlock",
-		},
-		permissions: [
-			{
-				id: "513",
-				controller: "userspermissions",
-				action: "getadvancedsettings",
-				enabled: true,
-			},
-			{
-				id: "1053",
-				controller: "activityassignee",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "1213",
-				controller: "category",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "1087",
-				controller: "contacttag",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "993",
-				controller: "organization-registration-type",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "327",
-				controller: "contenttypes",
-				action: "deletecontenttype",
-				enabled: true,
-			},
-			{
-				id: "755",
-				controller: "budget-targets-project",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "1141",
-				controller: "tag",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "767",
-				controller: "budget-tracking-lineitem",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "1099",
-				controller: "country",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "881",
-				controller: "financial-years-org",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "967",
-				controller: "impact-units-org",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "491",
-				controller: "user",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1121",
-				controller: "organization",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "903",
-				controller: "fund-receipt-project",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "287",
-				controller: "contentmanager",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "459",
-				controller: "proxy",
-				action: "uploadproxy",
-				enabled: true,
-			},
-			{
-				id: "2741",
-				controller: "deliverable-category-org",
-				action: "deliverable_category_achieved_target",
-				enabled: true,
-			},
-			{
-				id: "1043",
-				controller: "activity",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "1185",
-				controller: "impact-linitem-fy-donor",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "1067",
-				controller: "activitytype",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "1075",
-				controller: "contact",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "799",
-				controller: "deliverable-category-unit",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "1235",
-				controller: "unit",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "517",
-				controller: "userspermissions",
-				action: "getpermissions",
-				enabled: true,
-			},
-			{
-				id: "1017",
-				controller: "project",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "1027",
-				controller: "workspace",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "1135",
-				controller: "state",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "945",
-				controller: "impact-target-project",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "749",
-				controller: "budget-targets-project",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "769",
-				controller: "budget-tracking-lineitem",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "929",
-				controller: "impact-category-unit",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1049",
-				controller: "activityassignee",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "519",
-				controller: "userspermissions",
-				action: "getpolicies",
-				enabled: true,
-			},
-			{
-				id: "279",
-				controller: "components",
-				action: "listcomponents",
-				enabled: true,
-			},
-			{
-				id: "1149",
-				controller: "village",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "927",
-				controller: "impact-category-org",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1045",
-				controller: "activity",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1059",
-				controller: "activitytype",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "875",
-				controller: "financial-years-donor",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "741",
-				controller: "budget-targets-donor",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "935",
-				controller: "impact-category-unit",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "2723",
-				controller: "deliverable-target-project",
-				action: "deliverable_achieved",
-				enabled: true,
-			},
-			{
-				id: "1197",
-				controller: "category-unit",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1109",
-				controller: "district",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "469",
-				controller: "upload",
-				action: "getsettings",
-				enabled: true,
-			},
-			{
-				id: "977",
-				controller: "organization-currency",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "301",
-				controller: "contenttypes",
-				action: "findcontenttype",
-				enabled: true,
-			},
-			{
-				id: "1169",
-				controller: "financial-year",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1115",
-				controller: "district",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "735",
-				controller: "budget-category-organization",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1157",
-				controller: "village",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "2711",
-				controller: "impact-category-org",
-				action: "totalachivedimpactprojectbyorg",
-				enabled: true,
-			},
-			{
-				id: "1173",
-				controller: "deliverable-linitem-fy-donor",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "1165",
-				controller: "financial-year",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "757",
-				controller: "budget-targets-project",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "461",
-				controller: "upload",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "743",
-				controller: "budget-targets-donor",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "263",
-				controller: "public-donor",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1229",
-				controller: "sustainable-development-goals",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "1249",
-				controller: "deliverable-category-org",
-				action: "projectcountdelcatbyorg",
-				enabled: true,
-			},
-			{
-				id: "1233",
-				controller: "unit",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1199",
-				controller: "category-unit",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "797",
-				controller: "deliverable-category-unit",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1253",
-				controller: "impact-category-org",
-				action: "projectcountimpcatbyorg",
-				enabled: true,
-			},
-			{
-				id: "1089",
-				controller: "contacttag",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "283",
-				controller: "contentmanager",
-				action: "checkuidavailability",
-				enabled: true,
-			},
-			{
-				id: "1139",
-				controller: "tag",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "845",
-				controller: "deliverable-units-org",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "825",
-				controller: "deliverable-tracking-lineitem",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "729",
-				controller: "budget-category-organization",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "1195",
-				controller: "country",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "509",
-				controller: "userspermissions",
-				action: "deleteprovider",
-				enabled: true,
-			},
-			{
-				id: "1",
-				controller: "account",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "1025",
-				controller: "workspace",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "739",
-				controller: "budget-targets-donor",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "1031",
-				controller: "workspace",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "1029",
-				controller: "workspace",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "531",
-				controller: "userspermissions",
-				action: "init",
-				enabled: true,
-			},
-			{
-				id: "727",
-				controller: "budget-category-organization",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "827",
-				controller: "deliverable-tracking-lineitem",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "1243",
-				controller: "unit",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1207",
-				controller: "category-unit",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "1193",
-				controller: "impact-linitem-fy-donor",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "839",
-				controller: "deliverable-unit",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "791",
-				controller: "deliverable-category-org",
-				action: "find",
-				enabled: true,
-			},
-			{
-				id: "2732",
-				controller: "budget-targets-project",
-				action: "donors_recieved_value",
-				enabled: true,
-			},
-			{
-				id: "883",
-				controller: "financial-years-org",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "873",
-				controller: "financial-years-donor",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "889",
-				controller: "financial-years-org",
-				action: "findone",
-				enabled: true,
-			},
-			{
-				id: "963",
-				controller: "impact-tracking-lineitem",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "2714",
-				controller: "impact-category-org",
-				action: "totalimpactprojectbyorg",
-				enabled: true,
-			},
-			{
-				id: "289",
-				controller: "contentmanager",
-				action: "delete",
-				enabled: true,
-			},
-			{
-				id: "975",
-				controller: "impact-units-org",
-				action: "update",
-				enabled: true,
-			},
-			{
-				id: "463",
-				controller: "upload",
-				action: "destroy",
-				enabled: true,
-			},
-			{
-				id: "1083",
-				controller: "contacttag",
-				action: "create",
-				enabled: true,
-			},
-			{
-				id: "533",
-				controller: "userspermissions",
-				action: "searchusers",
-				enabled: true,
-			},
-			{
-				id: "511",
-				controller: "userspermissions",
-				action: "deleterole",
-				enabled: true,
-			},
-			{
-				id: "761",
-				controller: "budget-tracking-lineitem",
-				action: "count",
-				enabled: true,
-			},
-			{
-				id: "931",
-				controller: "impact-category-unit",
-				action: "create",
-				enabled: true,
-			},
-		],
-	},
-};
